possible null pointer
Bug: Possible null pointer dereference in
org.bertolo1988.imagifier.imagifier.Imagifier.buildSampleColorImages()
due to return value of called method

The return value from a method is dereferenced without a null check, and
the return value of that method is one that should generally be checked
for null. This may lead to a NullPointerException when the code is
executed.